+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 1.134 (not yet released)
 -------------
 
+* CreateContribution now also takes Financial Type as parameter (For the configuration Financial Type is still required but will function as default).
 * Subscribe to mailing list, improve error reporting and parameter checking, consistent parameter label. See !147
 
 Version 1.133